+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKBJ1G3667RPMG3YKKN887S4"
+        "01KKBPGCCTE7KM3MMYPC7H2VRN"
     )
     r"""The unique identifier of the data source"""